+++ Fun.Htmx/DslHtmxBuilder.fs
@@ -2,6 +2,7 @@
 module Fun.Htmx.DslHtmxBuilder
 
 open System.Text
+open Microsoft.AspNetCore.Http
 open Fun.Result
 open Fun.Blazor
 open Fun.Blazor.Operators
@@ -45,6 +46,16 @@ type DomAttrBuilder with
     member inline _.hxRequestBlazorSSR<'T>([<InlineIfLambda>] render: AttrRenderFragment, queryBuilder: QueryBuilder<'T>) =
         render ==> ("hx-get" => $"/fun-blazor-server-side-render-components/{typeof<'T>.FullName}?{queryBuilder.ToString()}")
 
+    /// Issues a get request to get the blazor component and reader as static dom
+    [<CustomOperation "hxGetComponent">]
+    member inline this.hxGetComponent<'T>([<InlineIfLambda>] render: AttrRenderFragment, queryBuilder: QueryBuilder<'T>) =
+        this.hxRequestBlazorSSR(render, HttpMethods.Get, queryBuilder)
+    
+    /// Issues a post request to get the blazor component and reader as static dom
+    [<CustomOperation "hxPostComponent">]
+    member inline this.hxPostComponent<'T>([<InlineIfLambda>] render: AttrRenderFragment, queryBuilder: QueryBuilder<'T>) =
+        this.hxRequestBlazorSSR(render, HttpMethods.Post, queryBuilder)
+
 
     /// Issues a request to get the blazor custom element as the return dom, 
     /// and it will open a websocket for the component's interactivity
@@ -66,6 +77,18 @@ type DomAttrBuilder with
     [<CustomOperation "hxRequestCustomElement">]
     member inline _.hxRequestCustomElement<'T>([<InlineIfLambda>] render: AttrRenderFragment, queryBuilder: QueryBuilder<'T>) =
         render ==> ("hx-get" => $"/fun-blazor-custom-elements/{typeof<'T>.FullName}?{queryBuilder.ToString()}")
+
+    /// Issues a get request to get the blazor custom element as the return dom, 
+    /// and it will open a websocket for the component's interactivity
+    [<CustomOperation "hxGetCustomElement">]
+    member inline this.hxGetCustomElement<'T>([<InlineIfLambda>] render: AttrRenderFragment, queryBuilder: QueryBuilder<'T>) =
+        this.hxRequestCustomElement(render, HttpMethods.Get, queryBuilder)
+        
+    /// Issues a post request to get the blazor custom element as the return dom, 
+    /// and it will open a websocket for the component's interactivity
+    [<CustomOperation "hxPostCustomElement">]
+    member inline this.hxPostCustomElement<'T>([<InlineIfLambda>] render: AttrRenderFragment, queryBuilder: QueryBuilder<'T>) =
+        this.hxRequestCustomElement(render, HttpMethods.Post, queryBuilder)
 #endif